CI: Add retries to ansible stages
On several occasions there have been failed CI runs that were aborted
because of strange and non-deterministic Ansible failures. These false
negatives (the change that is tested is actually fine) obviously harms
the confidence in the CI.

This commit adds retries to the single stages
and makes use of the fact that the playbooks are idempotent s.t. a weird
timeout might resolve itself on the second attempt. Reruns are usually
also alot faster so the additional waiting time to see one's MR fail
stays reasonable.