Create single & dualstack load-balanced svc checks
This commits adds tasks to "check-loadbalancer-service" which validate
the creation of single stack IPv4, single stack IPv6 and dual stack
load-balanced services.

The tasks use the newly introduced ipFamilies array and the
ipFamilyPolicy attribute for service definitions.

ch-k8s-lbaas does not support IPv6 single stack load-balanced services,
therefore we do disable this task block. f1a does not support IPv6 and
it is unclear how we want to handle/design the IPv6 stack and how to
handle the load-balancing in Yaook.

Issue: #227