k8s-config: deduplicate AS variables
Calico as well as kube-router use variables for the AS number of the
gateways and the AS number of the k8s nodes. However, the variables
were defined redundantly.

This commit deduplicates these variables. Additionally, this commit
moves calico related variables to the calico role. They have been
present at the k8s-config role so that they are grouped with the AS
variables. This is not necessary anymore.

Issue: #270