net-forwarding: introduce role for sysctl config
We need to enable forwarding on the gateways as well as on the k8s nodes
for both, IPv4 and IPv6 addresses (if DualStack is enabled). In advance
to this commit this has been done at multiple points. Redundant code
should be unraveled.

This commit introduces a new role "net-forwarding" which configures
the sysctl entry for IPv4 and IPv6. This newly introduced role has been
added as dependency to the corresponding roles and the tasks to
configure sysctl to enable forwarding have been removed from other roles.

The reason for introducing a baseline role is quite simply a better
repository overview.

Issue: #227