ipsec: create CHILD_SA per subnet in remote_ts
IKEv2 supports multiple subnets in the remote_ts field of a CHILD_SA.
For various and partially unknown reasons [0] the peer will only accept a
subset of these subnets (usually the first subnet). In consequence only
this subnet is addressable and traffic to the others is lost.

As a workaround, this commit creates a CHILD_SA per remote subnet in the
configuration. [1] See also: https://redmine.cloudandheat.com/issues/30472

[0] https://lists.strongswan.org/pipermail/users/2017-May/010996.html
[1] https://wiki.strongswan.org/projects/strongswan/wiki/FAQ#Multiple-subnets-per-SA