Use BGP for wireguard subnet route propagation
The previous setup had a lot of drawbacks. A virtual IP has been added
via terraform that has been managed by keepalived. This virtual IP has
been used to determine if radvd should start advertising a route to
the wireguard subnet.

We do not want an additional routing daemon and the virtual IP overhead.

The main problem here is, that wireguard runs on all gateways and each
gateway thinks that it knows a route to the wireguard peers, but only
the currently lb-master gateway knows (because the wg peer is connected
to it).

This commit adjusts the setup so that the route to the wireguard subnet
is propagated via BGP (bird) by the current lb-master gateway.

Depending on which gateway has the Virtual IPs for load-balancing, this
gateway will peer with the k8s masters. The wireguard role has been
extended to define an export filter for bird which includes the wireguard
subnet. This is important, because we need to define that in stage2,
but the peering with the k8s nodes is done in stage3. We need to ensure
the separation of the stages. On the k8s side, calico/node ensures that
the peered k8s node does import the route to the wireguard subnet.

Issue: #227