Peer lb-master-gateway with every k8s node
We need to peer all k8s nodes with the currently active load-balanced
master gateway, because calico/node only forwards routes (via BGP)
belonging to an IPPool. As we want to propagate how the nodes can reach
the wireguard subnet, we need to peer them directly with the gateway.

This is necessary to directly talk to k8s nodes from outside the cluster.

Adding the wireguard to a disabled IPPool would prevent this necessity,
but is not possibly because the wireguard subnet is not known to stage3.

Besides that, this commit cleans up the BGP setup procedure of calico.

Issue: #227