Add proper semver comparison for terraform
Since !194, Terraform versions < 0.14.0 are not supported anymore.
Thus, the MR added a simple version comparison mechanism which solely
relied on comparing the minor portion of the string. With the release
of Terraform v1.0.0 on 2021-06-08, this doesn't work anymore.

This commit adds a proper version comparison that will yield the
correct result as long as the version adheres to the semver spec.