Make threshold for pod garbage collector configurable
Default is terminated_pod_gc_threshold=10

The Kubernetes garbage collector deletes objects that don't have an
owner anymore. According to the pentest report, the current setting for
garbage collection is 12,500 terminated pods which is probably too high.

Note that once a terminated pod is cleaned up, its logs are no longer
accessible.

In context of #282