rook: Move v1.2 manifests to dedicated directory
As we want to support more and future rook versions and the manifests,
slightly differ from version to version, the manifests for
v1.2 have been moved to a dedicated directory.
The main.yaml has been adjusted to use the manifests of the currently
configured (major.minor) version (in k8s-config).
Comments have been added to the manifests to explain some parameters.

Additionally, this commit adds parameters which are necessary for the
upgrade to v1.3. If we have no redundancy for OSDs defined, we need to
enable "skipUpgradeChecks" in the cluster.yaml. However, we will
add cluster health verification tasks in the following commits.
When the pool size is one, "requireSafeReplicaSize" should be disabled.
Otherwise the rook-ceph-operator will refuse to perform an update on
the deployment.The same applies to the filesystem.

Issue: #123