rook: Add rook/ceph version compliance checks
Before performing upgrade steps or applying manifests, it is necessary
to validate that the configured rook and ceph versions (in k8s-config)
are a) supported by us and b) the ceph version is supported by the rook
version.

This commit adds an array of dicts to the role variables where future
versions have to be added after support has been added to the role.
In addition to the rook version which is supported, the major ceph
version which this rook version supports have to be defined.

That the configured rook and ceph versions are valid is checked in
newly added tasks (in advance before any change is made to the cluster).

Issue: #123