rook: Add ceph upgrade tasks
Upgrading ceph is as simple as patching the image version of the
CephCluster resource. The rook-ceph-operator will update all components
(in the rook context these are all deployments) to use the newly
configured ceph version.

This commits adds three steps to the k8s-rook role. First, the version
checks have been extended to check if ceph needs to be upgraded (if a
new version has been configured). Second, the actual step of patching
the ceph version. And thirdly, after the ceph version has been patched,
the rook-ceph-operator will take some time to perform the update.
Since a possibly followed rook upgrade needs a healthy ceph cluster,
we wait and check that all rook deployments are updated und use the
newly defined ceph version.

To prevent spamming, we pause the run for 3 minutes and watch in
amazement as the cluster updates. After the 3 minutes we start checking
if the upgrade has been finished.

Issue: #123