rook: Verify that rook version upgrade is allowed
This commit adds checks that validate if the configured rook version
is allowed to be deployed. Or in other words: if the updated is allowed.

Rook can only be updated step by step from minor to minor version.
This makes the version checks a bit complex.
We do not care about patch versions. Patch versions are applied by
simply patching the rook-ceph-operator image.

Issue: #123