rook: Add manifests for version v1.3
This commit adds manifest templates for rook v1.3 to a dedicated
directory. These manifests will be applied automatically if a v1.3
patch version has been configured in k8s-config.

The volumeClaimTemplate for the StorageClassDeviceSet "cinder" has been
renamed to "data" (from "ceph-data"). Otherwise, OSDs will not come up,
because the rook-ceph-operator complains that the data template is
missing. In the case of an update, the rook-ceph-operator won't update
the OSD deployments.

Issue: #123