rook: Add rook upgrade verification
After the actual upgrade tasks have been executed, the rook-ceph-operator
will take some time to update all the rook deployments.

To prevent spamming, we pause the run for 3 minutes and watch in
amazement as the cluster updates. After the 3 minutes we start checking
if the upgrade has been finished.

This commits adds tasks which will wait and validate that all rook
deployments have been updated to use the newly configured rook version.

Issue: #123