rook: Make operator upgrade checks configurable
This commits adds variables for the cluster manifest parameter
"skipUpgradeChecks" and variables for "requireSafeReplicas" for the
filesystem and pool manifests.
These variables have been added to the config template and have been
defined for the CI config.

If the pool replicas are 1 (they're currently by default), the rook-ceph-
operator will not perform update on these deployments. Same applies for
the filesystem. Therefore, we'll set "requireSafeReplicas" automatically
based on the configured replica size.

In addition, it is possibly to configure that the rook-ceph-operator
skips upgrade checks before updating the ceph cluster. This is not as
bad as it sounds, because we added cluster health verification tasks
in advance.

Starting with rook v1.3, the cluster will be in
health warning state if the OSDs are not redundantly rolled out.
Skipping upgrade checks is therefore necessary for CI alike clusters.

If you build bigger, redundant ceph clusters with rook, unsafe replica
sizes should not be configured and so that the rook-ceph-operator will
perform upgrade checks.

Issue: #123