rook: Add v1.4 upgrade tasks
This commit adds tasks that perform the actual update from rook v1.2
to rook v1.3.

Similar to the upgrade v1.2 to v1.3, the RBACs are updated, the CRDs
are updated and the rook-ceph-operator is updated. We do not need to
apply the new CephCluster manifest when upgrading from v1.3 to v1.4.

Issue: #292