rook: Add manifests for version v1.6
This commit adds manifest templates for rook v1.6 to a dedicated
directory. These manifests will be applied automatically if a v1.6
patch version has been configured in k8s-config.

The manifests are kept as close as possible to the templates of the rook
repository.

Since rook v1.6 we need to set "continueUpgradeAfterChecksEvenIfNotHealthy"
along with "skipUpgradeChecks". Otherwise OSDs will not be updated.

I had to remove the creation timestamps in the CRDs. Otherwise
reapplying them fails.

Issue: #295