rook: Add config for rook version, add version map
This commit introduces the possibility to configure the rook version
to be deployed in the cluster config toml file. A rook-ceph version map
has been added to k8s-config that will map the configured rook version
to a ceph version. The default is rook v1.2.3 with ceph v14.2.5, because
this way we preserve compatibility with existing clusters. The config
option has been added to the config template as well as to the ci config.

Hence, a rook/ceph upgrade can now be configured by simply adjusting
the rook version in the configuration file.

Issue: #123