rook: Update ceph cluster subsequent to rook
By upgrading ceph subsequent to the rook-ceph-operator (and other rook
related deployments) we ensure that we can jump directly to the latest
supported ceph version.

If e.g. a new minor release of rook supports ceph v16, we could not
directly jump to this version because the rook-ceph-operator would refuse
to perform the update, because the old rook version is still used and
the operator will refuse to perform an updated to an unsupported ceph
version.

By updating the rook-ceph-operator prior to the ceph cluster, we can
update to the latest rook release as well as to the latest ceph release
in one execution.

Issue: #123