cloud-config: Limit OpenStack network name
The OpenStack cloud-controller-manager will, by default, add node ips
into all networks to which a node is attached. Often this is not what
you want because your CNI plugin might, by chance, pick-up the wrong one
and mess up your cluster network.

This commit limits the selection of the CCM to the listed network.