Fix networkpolicy enforcement check
We are experiencing CI job failures for mk8s clusters with calico as
CNI (and rook, but I highly doubt that that's related). This may be
introduced by resizing the CI runner instance (which does not really
make any sense).

Firstly, this commit fixes a bug in the check that the service is not
reachable from the node. The "-w" option was followed by "--connection-
timeout" this lead to the fact, that instead of the HTTP code, the
timeout flag has been printed out. Additionally, this lead to the fact
that the "--connection-timeout" flag has been ignored. Unlike kuberouter,
calico does not send a connection refused if a network policy rejects
access to the service. Instead the connection runs into a timeout.
I do not know which exact timeout is used, but my guess is about 5m.

My guess is, that the CI somehow is confused / errors because the specific
task takes so long. This is probably also related to the dockerization.

Because we do not want to wait for the default calico timeout and the
timeout flag of curl is now respected, this commit adds "28" to the
accepted rc. Instead of retrying via ansible, this commit adjusts the
task such that curl is retrying directly. This is due to the fact that
we want to make sure that the connection timeout has not happened by
accident. Otherwise the task would succeed after 1 try. Now 3 retries
are forced.