Fix k8s node labels and taints
The k8s node specific labels and taints are host specific. Previously,
they were not exported correctly by "03_z_export_config", because
they are not present in the top layer of the vars. As the labels and
taints were not exported correctly, they could not be applied by the
k8s-service-layer. This lead to the fact that e.g. the rook-ceph-operator
or monitoring components stayed pending and the ansible playbook ran
into a timeout.

Exporting the labels and taints requires iterating over all hostvars.
This commit adjusts the export of the config for the service layer
such that labels and taints are exported for each host / k8s node.
In addition, the appliance of the labels in the scheduling_info_v1
playbook has been adjusted to meet the requirements of the newly
introduced nodes_labels_taints-Array of the config.

Besides that, the config export has been adjusted such that the useless
key-value pair which was there to facilitate meeting the requirements of
the JSON format can be removed.