Automatically scale the typha deployment
Calico Typha is tricky to rollout/restart. It uses host networking while
still exposing ports, which means that a new replica cannot be scheduled
on a host where an old replica is already running.

This means that we can only have N typha replicas, where N is the
number of nodes on which typha could possibly run. We need to instruct
the Deployment to take this into consideration, by allowing it to kill a
typha in order to start new ones.

The documentation [1] states that "one Typha instance can support
hundreds of Felix instances", which means that we don't need three
typhas in our clusters typically.

The new logic is extensively explained in the deployment template, so
feel free to read there for more details.

   [1]: https://docs.projectcalico.org/reference/typha/overview