Create frontend group in Terraform
This is an extension hook to allow LCMs building on top of
managed-k8s to provide their own set of nodes which handle the
frontend tasks. This is generally out of scope for the
terraform-based setup, however.

The prime user will be the Yaook project where the inventory
(and nodes) is not generated (provisioned) via Terraform, but via
other means.

Future changes will build on top of this to fade out certain uses
of the `gateways` group with the `frontend` group for that
purpose.