Move choice of Python interpreter to lower layer
In general, we want to default to an interpreter which is
Python 3. However, on bare metal CentOS 8, we are having some fun
with the choice of python3 vs. platform-python. Packages are only
available for platform-python, but if a newer python version is
installed, python3 will point at that instead of platform-python.

This means that crucial things such as dnf are not invokable from
the non-platform-python interpreter.

Unfortunately, ansible 2.9 has no sane default fallback path for
the python version on Debian and Ubuntu [1], which means that we
cannot simply rely on Ansible's auto-detection of the
interpreter for our normal Managed Kubernetes clusters.

In order to resolve this conflict, we relegate the responsibility
for choosing a python interpreter to the provisioner layer
(Terraform or Yaook Bare Metal); they will generally know better
which interpreter to choose anyway.

   [1]: https://docs.ansible.com/ansible/2.9/reference_appendices/config.html#interpreter-python