tf: Enable anti-affinity group for workers
Server groups are a feature offered by OpenStack Nova by which the user
can influence the scheduler on where to put machines. Common policies
are "affinity" (keep instances on one host) and "anti-affinity" (spread
them apart). There're also weaker forms that are prefixed with a "soft"
that work on best effort.

This commit

- creates a default server group with an anti-affinity policy and
- provides a list to which the user can add instances that must not be
  placed on the same hypervisor (UX similar to specifying flavors or
  images)

The extension is intentionally kept so narrow-scoped (e.g., not
offering the same option for gateways and or masters) because its
usecase is so limited. (https://redmine.cloudandheat.com/issues/31058)

Existing VMs aren't touched (touch = destroy + recreate) so the initial
setup has to be done by hand.