Split calico role into control-plane and worker parts
When deploying a cluster without a central deployment coordinator
node (e.g. in the Yaook bare metal -> kubernetes case), it is
desirable to have a clean separation between tasks executed on
control plane nodes and tasks executed on worker nodes.

The current calico plays had three category of tasks:

- delegated to localhost (to generate secrets)
- using calicoctl/kubectl with the admin.conf
- normal tasks directly executed on the target host

The first two categories only need to run once, generally, and they
require privileges or a level of trust. The third category can run
autonomously on any node on its own.

In order to facilitate autonomous bootstrapping of a node, the tasks
which run on each node independently are factored into the calico-worker
role, while the other tasks are grouped into calico-control-plane. Those
are then called from the respective master/worker sub-plays in the
03_final play. This implies that after those plays have been run, we
have a cluster which has working networking.

Note that we cannot do the calico-control-plane role in the same step as
the k8s-master role, because k8s-master is deployed using `serial`,
while we need to run calico-control-plane only after all masters
currently defined in the inventory are up and clustered in order for the
typha rollout to work (otherwise, scheduling will not allow typha to
come up completely, which makes the run fail).

In context of the attempt in !276, this would allow us to avoid the ugly
tag-based hack and instead have a clean top-level point in the play
where we can safely exit.