Disable monitoring checks
We have the issue that one Node-Exporter Pod is pending when
initializing a fresh CI-alike cluster. It cannot be scheduled on one
storage node due to insufficient memory.

However, after draining and uncordoning all nodes, the issue is resolved
and everything is up and running. So we may want to
a) figure out why the heck this is happening now
b) simply use a bigger flavor for storage nodes
c) drain/uncordon all nodes to finish a fresh cluster init

I'll leave that question to future me and future you