Refactor the inventory handling (toml helper)
This commit refactors how we are handling ansible inventories as well
as necessary variables for e.g. Terraform. In the previous setup,
all the logic was contained in the config.toml. This had the advantage
that the logic for the toml_helper was relatively simple.
However, that is not the full truth. The toml_helper actually was a mess.
It did things in an unstructured way and it was really hard to comprehend
what it is actually doing.

The config was designed in a way that the logic where a variable will be
used (in which layer) was in the format of the config itself. This made
the correct configuration complicated and required the developer to have
non-negligible prior knowledge of the entire setup.
In addition, this structure meant that variables used in several levels
had to be either defined twice or moved to general inventory (refer
ansible/group_vars/all/config.yaml). The swapping made it very opaque to
recognize where the variable value is actually defined concretely.

This commit tries to uncordon and resolve these things. Helper scripts
(or rather modules) for wireguard, terraform and pass have been created.
This allowed to push the whole logic regarding these services to be
pushed into separate modules. Thereby, the actual inventory handler
can be kept a bit more generic. The logic where to put a variable and
how to group variables is now mainly contained in the inventory helper.

The inventory handler is now named "update_inventory" and
is contained in the "actions" directory. The helper scripts for the
inventory handler reside in a "helpers" subdirectory.

All resources were moved out of `jenkins/` (either to templates/ or to action/).
Also, the name was extremely misleading because we never used Jenkins
in the context of mk8s.