Add legacy config converter
This commit introduces a legacy config converter as part of the
refactored inventory handler. We need a legacy config converter, because
we need to automatically translate existing cluster (especially
productive one) into the new format.

The legacy config converter is based upon the previous config template.
It uses a map which defines variables we expect to see. As the
configuration has been restructured, the converter needs to some pushing
around of sections and variables.

If the converter cannot process all variables, it will fail and you'll
need to manually check and fix your configuration. Afterwards you can
retry again.

The refactored inventory handling requires some new variables to be
defined in the config, because e.g. their default values are not
present in-line the code anymore. Therefore, after translating the
legacy config we do merge it with the template such that these variables
are present in the config without the need of especially adding them.

After you manually confirmed that the new config looks promising, it
will be passed back to the inventory handler which will then normally
proceed.