Move global monitoring stuff from stage2 to stage3
This commit introduces a new role "global-monitoring-support" which
adjusts HAProxy and keepalived for the global monitoring. This role is
executed in stage3. Previously, the configuration happened in stage2.
Moving it to stage3 has the advantage that we can leave out global
monitoring variables for stage2 which makes inventory handling simpler.