Redo variable handover for KMS and KSL
The previous approach was unsatisfying. Also it was not sure if it
preserved whitespaces and quotations correctly.

This commit refactors the variable handover by using a combination of
the "pre_tasks" feature and the "include_vars" ansible module.
This ensures that the stage specific variables are available.

We could even refine the used variables per role/task now by adjusting
the pre_tasks for each role and including only variables that are
really used by the role. For now, giving each role access to all
stage-specific variables is the simplest solution and I'll leave the
refinement open for future generations.