Fix rook scheduling (revert changes)
This commit reverts the previous changes made to the rook manifests
regarding the rook_nodeplugin_toleration and the scheduling_key.

I confused how these vars are actually intended to be used. The
rook_scheduling_key is a string which controls where rook services
should run. The rook_nodeplugin_toleration can either be "false",
"true" or an arbitrary string "foo". If it is true, then everything
is going to be tolerated. If it is false, only the defaults will be
tolerated. If it's a string, then this is the string to be tolerated.