config.template: Comment hostname-specific fields
config.template.toml serves both as a an initial configuration file for
new clusters as well as an additional source of fields that is merged
into the config the legacy converter creates.

This commit removes to-be-merged fields that have a hostname-specific
component in them because we cannot assume that, e.g., "managed-k8s-worker-5"
exists in every environment. Such assumptions are not only misleading
but also cause some roles to fail.