Ensure that control plane nodes peer with each other
The previous BGPPeer configurations did not allow route-reflector nodes
to peer with one another. It appears that an empty peerSelector does
nothing instead of every node.

This reverts to the state before e864ed527, where there was an explicit
BGPPeer configuration which said "all nodes are to peer with route
reflectors". This creates a working cluster.

Without that, non-host-network pods on the control plane nodes were not
able to communicate with one another or with cluster services, including
crucial services such as coredns, if they were not hosted on the same
control plane node.

That broke, among other things, the cinder CSI plugin (but only if it
was scheduled on a node which did not run coredns).

This is a partial revert of e864ed527.