Fix forwarding/router advertisment config
All of the explanation is inline really. No need to replicate that in
this commit message I guess.

The effect which was observed was that after a day or so (the lifetime
of the router advertisments sent by the OpenStack routers), the gateways
would lose all IPv6 connectivity: they lose the /120 to their neighbours
as well as the default route. That was because forwarding was turned on
in the meantime, which makes Linux assume that you're not interested in
RAs at all.