Fix wireguard helper IP address collision
Existing/predefined addresses (in the config) are handled as
IPv<4,6>Interface objects by the wireguard_helper. This is because
the fromdict function returns the address as interface object.

This lead to the fact that a collision between predefined addresses
(like e.g. the ones for the gitlab CI runner) and newly allocated
addresses were not detected immediately.
Two users may end up with the same IP which will later lead to a
failure of the run [1].

This commit adjusts the wireguard_helper such that the fromdict function
returns the address as actual ip address object so that not
an Interface object is used in later procedures.

This does not lead to any changes in the address format in the config
file. It is possible to predefine an address with or without prefix.
However, in the wireguard_ipam file, the prefix will be removed.

[1] https://gitlab.cloudandheat.com/lcm/managed-k8s/-/jobs/192248