ci: fetch images from gitlab.com registry
Note 1: the tag `f1a:deel` is wrong but cannot be changed right now
because the author of this commit does not have the necessary
permissions in https://gitlab.com/yaook/images/k8s-ci.

Note 2: other yaook projects use the special environment variable
`CI_DEPENDENCY_PROXY_GROUP_IMAGE_PREFIX`. For the sake of getting the CI
to work as quickly as possible the use of the caching proxy has been
postponed.