Remove ch-role-users submodule
ch-role-users is a C&H company internal module that must not be exposed
to the public. Unfortunately it's tightly coupled into the LCM
repository. This commit removes it as a submodule so that it is not
automatically fetched via `git submodule update --init --recursive`.
In the upcoming commits the rest of the LCM will be adapted to cope with
the absence of the role.