Create symlink to ch-role-users
As discussed in !323, making the submodule / ch-role-users repository
opt-in caused us headaches. We decided to move the ch-role-users
submodule to the cluster repository. This is consistent with the other
company internal repositories (wireguard users, passwordstore users).

This commit adds a symlink to the ch-role-users submodule in the cluster
repository. For non-C&H users this link will be broken. However, that
should be no problem as the role is only included if configured so (via
env var or configuration file).