CI: Fix ansible linting stage
This commit ensures that the ansible-linting stage (CI) actually checks
all roles and playbooks. To do so, this commit adjusts and moves the
ansible-lint configuration file to "ci/lint/". The configuration file
is adjusted such that all roles and playbooks get checked (one after
another).

I tried to simply run ansible-lint against the base directory and
additionally add

kinds:
  - playbook: "ansible/02_trampoline.yaml"          # Check stage 2
  - playbook: "ansible/03_k8s_base.yaml"            # Check stage 3
  - playbook: "k8s-service-layer/install.yaml"      # Check stage 4
  - playbook: "k8s-managed-services/install.yaml"   # Check stage 5

to the configuration file.

Ansible-lint will then search for all roles/playbooks and additionally
examine the configured playbooks again. This approach failed!
Ansible-lint will fail when examining the ch-role-users role if it is
not available. It is also possible to ignore paths/files, however this
seems not to be the case when ansible-lint is examining a role not in
the context of a playbook.

Adjusting the ansible linting configuration file to ignore the
ch-role-users role is not working. At least I was not able to get it
working. Among other things I tried the following configuration:

exclude_paths:
  - "./ansible/roles/ch-role-users"   # ignore C&H internal role
warn_list:
  - ch-role-users
  - load-failure

Removing the symlink is a working solution, but feels wrong. Instead,
we are going to trigger ansible-lint against all stages and test phases
explicitly.