CI: Adjust size and number of monitoring nodes
Previously we had 3 monitoring nodes of size "M". With this setup it was
not possible to create a managed-k8s cluster with monitoring and rook
enabled. The "prometheus-k8s" Pod will be pending and the monitoring
checks will fail. (Due to insufficient CPU available in combination
with labels/taints).

This commit adjusts the number and size of monitoring nodes so that we
now use two (2) "L"-nodes for monitoring. This will ensure that all Pods
can be scheduled. An additional (sixth) worker node with the flavor "L"
is available to ensure that all all tests are able to be executed. E.g.
for  the "check-shared-fs" test we want to have three different nodes at
least.

The quota of the CI project has been adjusted to meet these requirements.

A drawback is that most developer projects have oriented on the CI
quota and probably will need to adjust their quotas of their developer
projects as well if they want to enable monitoring and rook.

Issue: #334