Monitoring-checks: Simplify node exporter scraping
The checks are immediately executed after the monitoring setup. It can
(but not must) happen that the endpoints are not up yet. A retry-until
logic has been added to the tasks. This commit simplifies and fixes
the retry-until logic. The previous one did only work if the node
exporter were already up, but the task failed after the first try if
they weren't. Probably the node-exporters delivered too late and
prometheus checked too early.

Issue: #334