CI: Fix duplicated pipelines and minor adjustments
This commit adds a workflow definition such that a pipeline is only
created for merge requests, schedules, and direct commits to devel.

This commit adjusts the linting jobs such that the rules key is obsolete.
This is just for better comprehensibility.

This commit moves duplicated keys of the spawn cluster jobs to the
common configuration like e.g. the image reference and tags. This does
also improve comprehensibility.

Related:
https://docs.gitlab.com/ee/ci/jobs/job_control.html#avoid-duplicate-pipelines

Issue: #331