Fix use of gateways vs. frontend in keepalived config
This fixes a regression introduced in the DualStack merge request.
Since 0019a50, we use the `frontend` group to define the nodes which are
responsible for keepalived/haproxy. The DualStack MR did not account for
that and this causes issues in a deployment with only frontend nodes and
no gateway nodes (bare metal).