CI: Restrict triggering spawn-cluster jobs
Previously, the spawn-cluster jobs (at least the two necessary ones)
had to be executed before a merge request could be merged. After merging
the merge request, these were executed again due to the merge request
commit. We want to avoid duplicated execution and enabled merge trains [1].

However, simply enabling merge trains did not solve the actual problem.
Despite making the rebases unnecessary, the jobs were still executed
twice.

This commit adjusts the CI job triggers such that the jobs are only
triggered when an actual merge is happening, so inside the merge train.
If one of the job fails, the merge train will refuse to merge the mr.

Drawbacks: we cannot trigger the CI jobs without triggering a merge.
As probably not every contributor has a working test environment, he
has no feedback if the changes actually work in advance of the actual
merge. However, as the CI jobs are only triggered automatically or by
maintainers, we do not consider this a real problem.

On the long run, we need more CI environments.

[1] https://docs.gitlab.com/ee/ci/pipelines/merge_trains.html