check-storage-services: include rook defaults
By including the default values of the rook role we avoid defining
inline defaults for rook variables (especially rook_ceph_fs needs to
be known, otherwise the smoke tests will fail). Another advantage is
that we do not need to define the default value at multiple locations
with this approach.