Replace hard-coded nodeSelector in check-storage roles
For check-local-storage writer pod, use nodeSelector only when test_worker_nodes are defined.
For check-local-storage reader pod, use nodeAffinity to ensure that the same node as for writer pod is chosen.

For check-block-storage writer pod, use nodeSelector only when test_worker_nodes are defined.
For check-block-storage reader pod, use nodeAffinity to prefer the same node as for writer pod.

Signed-off-by: Roman Hros <roman.hros@dnation.cloud>