CI: lower rook-ceph-mon memory requests
tl;dr:
lower rook-ceph-mon memory requests to ensure that all node-exporters
can be scheduled (especially on storage nodes)

Currently, we have the issue that CI pipelines sometimes fail because
a node-exporter Pod cannot cannot be scheduled on a storage node due to
a lack of resources on that storage node despite the fact that the node-
exporter requires very few resources. As the node-exporter Pod cannot be
scheduled, the monitoring checks which try to scrape that endpoint will
fail. This problem happens in roughly 1/5 spawn-cluster jobs.

This one is hard to tackle as this never happened locally to me and it's
currently hard to debug because the artifacts are not descriptive enough
(refer #352). My guess was, that this happens when one rook metadataserver,
the ceph-operator, and one cephfs Pod (rook) along with the
csi-cinder-controllerplugin Pod are scheduled on the same storage node.
And this may happen because the rook-ceph-operator starts deploying
before the k8s-base (stage3) is completely up and settled.

HOWEVER: it has turned out my guess is wrong. I've seen a cluster job
failing where my assumption was not met.

This issue may be easily be fixed by using the Pod priority feature [1].
We currently do not support priority classes and adding this feature to
existing clusters will cause havoc as all Pods will have a priority of 0
(which is the lowest) until they are restarted.

Trying to tinker around with the kube-scheduler also feels wrong as it
basically does a pretty decent job. Also this feels like overengineering
this issue.

Deploying the monitoring in advance of rook is also not possible because
the monitoring is implicitly depending on rook.

This commit tackles this issue by adjusting the memory request for
rook-ceph-mon Pods. The memory limit will stay the same. The mons do not
complain that 500Mi is not sufficient (in contrast to the other rook
services that are already using the lower limit and currently cannot be
reduced any further).

[1] https://kubernetes.io/docs/concepts/scheduling-eviction/pod-priority-preemption/#priorityclass