Move system update monitoring to gateways only
The frontend nodes may be CentOS. The system update monitoring depends
on prometheus-node-exporter. Currently, our plays do not allow
installation of p-n-e on CentOS, which is why it makes no sense to have
the system update monitoring installed there.

In addition, there is a weird cross-dependency for the default of the
`monitoring_node_exporter_textfile_collector` variable which makes my
head hurt and which should hopefully be resolved by this.