Make Pod limit on k8s worker configurable
This commit adds support for configuring the maximum number of Pods
on each k8s worker.

End users may use a lot of small / tiny services and reach the default
Pod limit of 110 although the node is not even close to being fully
utilized. Normally, the host nodes have large RAM and CPU capacities
which increases the difference between Pod limit and node utilization.

The maximum number of Pods can be adjusted by changing the configuration
for kubelet. The default configuration file is located at:
/var/lib/kubelet/config.yaml. Changes to the configuration require a
restart of kubelet (a reload is not sufficient).

We are using "kubeadm" as bootstrapping tool. Currently it is possible
to configure kubelet via kubeadm during cluster initialization. However,
it is not possible to configure kubelet via kubeadm during a node join !

Therefore, a dedicated role "kubelet-configuration" is introduced by
this commit which can be used during various plays to adjust the
kubelet configuration on a host and restart the kubelet daemon.

This role has been integrated into the stage3 plays (it can be used in
both, the k8s cluster initialization as well as on a rerun) and the
upgrade scripts. During an k8s upgrade, kubelet must be restarted on
each host anyway, so that's a good opportunity to rollout the changes.
During a k8s cluster initialization we can also restart kubelet without
further consideration, because the k8s cluster is not up and running, yet.
On a rerun the rollout is guarded and skipped of disruptions are not
explicitly allowed.